perf: refactor benchmark infrastructure and unify test structure
Split monolithic utils.py into a package (env, results, runner, scene,
plotting, export), replace env-var config with pytest CLI options, separate
compilation from execution via jax.jit().lower().compile(), add early-stopping
via EnergyThresholdCondition for the empty-box benchmark, and unify both test
files (empty_box, directional_coupler) with matching _setup/_compile/_run
structure parametrized on plain ints with coarse/med/fine marks.
