+++ pyproject.toml
@@ -87,3 +87,8 @@ python-version = "3.14"
 
 [tool.hatch.metadata]
 allow-direct-references = true
+
+[dependency-groups]
+dev = [
+    "tensorboard>=2.21.0",
+]
