+++ tests/performance/conftest.py
@@ -1,19 +1,22 @@
 """Pytest configuration for performance tests.
 
-All tests in this directory are automatically marked as performance tests.
-Run only performance tests with:  pytest -m performance
-Exclude performance tests with:   pytest -m "not performance"
+All tests in this directory are automatically marked ``performance``.
+Run only performance tests:   pytest -m performance
+Exclude performance tests:    pytest -m "not performance"
 
 Each session writes to its own folder:
   tests/performance/results/{timestamp}_{device}/
     results.json      — benchmark timings and environment
-    figs/             — plots generated by the session
-    traces/           — XLA/Perfetto traces (PERF_TRACE=1)
+    figs/             — setup/material/field plots
+    traces/           — XLA/Perfetto traces (disable with --no-perf-trace)
+
+Tracing, memory profiling, and visualization are ON by default.
+Pass --no-perf-trace / --no-perf-memory / --no-perf-visualize to disable.
 """
 
 from __future__ import annotations
 
-import json
+import dataclasses
 import os
 
 # tests/conftest.py forces JAX_PLATFORMS=cpu to avoid sharding-mesh mismatches
@@ -25,15 +28,101 @@ from pathlib import Path
 
 import pytest
 
-from .utils import BenchmarkResult, capture_env
+from .utils import BenchmarkResult, capture_env, export_results
 
 _RESULTS_DIR = Path(__file__).parent / "results"
 
 
-def pytest_collection_modifyitems(items):
+# ---------------------------------------------------------------------------
+# CLI options
+# ---------------------------------------------------------------------------
+
+
+def pytest_addoption(parser: pytest.Parser) -> None:
+    g = parser.getgroup("performance", "FDTDX performance benchmark options")
+    g.addoption(
+        "--no-perf-trace",
+        action="store_false",
+        dest="perf_trace",
+        default=True,
+        help="Disable XLA/Perfetto tracing (enabled by default).",
+    )
+    g.addoption(
+        "--no-perf-memory",
+        action="store_false",
+        dest="perf_memory",
+        default=True,
+        help="Disable peak memory recording (enabled by default).",
+    )
+    g.addoption(
+        "--no-perf-visualize",
+        action="store_false",
+        dest="perf_visualize",
+        default=True,
+        help="Disable field intensity and setup plots (enabled by default).",
+    )
+    g.addoption(
+        "--perf-reps",
+        type=int,
+        default=1,
+        metavar="N",
+        help="Number of timed execution repetitions (default: 1).",
+    )
+    g.addoption(
+        "--perf-med",
+        action="store_true",
+        default=False,
+        help="Include medium-resolution variants of all benchmarks.",
+    )
+    g.addoption(
+        "--perf-fine",
+        action="store_true",
+        default=False,
+        help="Include fine-resolution variants of all benchmarks.",
+    )
+
+
+# ---------------------------------------------------------------------------
+# Collection hooks
+# ---------------------------------------------------------------------------
+
+
+def pytest_collection_modifyitems(config: pytest.Config, items: list[pytest.Item]) -> None:
     for item in items:
         if "performance" in Path(str(item.fspath)).parts:
             item.add_marker(pytest.mark.performance)
+
+    skip_med = pytest.mark.skip(reason="pass --perf-med to include medium-resolution variants")
+    skip_fine = pytest.mark.skip(reason="pass --perf-fine to include fine-resolution variants")
+
+    for item in items:
+        if item.get_closest_marker("perf_med") and not config.getoption("--perf-med"):
+            item.add_marker(skip_med)
+        if item.get_closest_marker("perf_fine") and not config.getoption("--perf-fine"):
+            item.add_marker(skip_fine)
+
+
+# ---------------------------------------------------------------------------
+# Fixtures
+# ---------------------------------------------------------------------------
+
+
+@dataclasses.dataclass(frozen=True)
+class PerfOptions:
+    trace: bool
+    memory: bool
+    visualize: bool
+    reps: int
+
+
+@pytest.fixture(scope="session")
+def perf_options(request: pytest.FixtureRequest) -> PerfOptions:
+    return PerfOptions(
+        trace=request.config.getoption("perf_trace"),
+        memory=request.config.getoption("perf_memory"),
+        visualize=request.config.getoption("perf_visualize"),
+        reps=request.config.getoption("--perf-reps"),
+    )
 
 
 @pytest.fixture(scope="session")
@@ -42,18 +131,11 @@ def perf_env() -> dict:
 
 
 @pytest.fixture(scope="session")
-def perf_run_dir(perf_env) -> Path:
-    """Create and return the per-session output folder.
-
-    All session outputs (JSON, figs, traces) land here so the entire
-    ``results/`` tree can be gitignored with a single rule.
-    """
+def perf_run_dir(perf_env: dict) -> Path:
+    """Create and return the per-session output folder."""
     ts = datetime.now(timezone.utc).strftime("%Y%m%dT%H%M%SZ")
     devices = perf_env.get("devices", [])
-    device_slug = (
-        devices[0].get("kind", "cpu").replace(" ", "_").lower()
-        if devices else "cpu"
-    )
+    device_slug = devices[0].get("kind", "cpu").replace(" ", "_").lower() if devices else "cpu"
     run_dir = _RESULTS_DIR / f"{ts}_{device_slug}"
     run_dir.mkdir(parents=True, exist_ok=True)
     (run_dir / "figs").mkdir(exist_ok=True)
@@ -63,32 +145,18 @@ def perf_run_dir(perf_env) -> Path:
 
 @pytest.fixture(scope="session")
 def perf_sink() -> list[BenchmarkResult]:
-    """Session-scoped list; tests append BenchmarkResult objects here."""
+    """Accumulates BenchmarkResult objects across the session."""
     return []
 
 
-def pytest_sessionfinish(session, exitstatus):
-    """Write collected benchmark results into the per-run folder."""
+def pytest_sessionfinish(session: pytest.Session, exitstatus: int) -> None:
     sink: list[BenchmarkResult] = getattr(session, "_perf_sink", None) or []
     run_dir: Path | None = getattr(session, "_perf_run_dir", None)
-
-    if not sink or run_dir is None:
-        return
-
-    ts = run_dir.name.split("_")[0]
-    env = sink[0].env if sink else {}
-    payload = {
-        "session_timestamp": ts,
-        "env": env,
-        "results": [r.to_dict() for r in sink],
-    }
-    out_path = run_dir / "results.json"
-    out_path.write_text(json.dumps(payload, indent=2))
-    print(f"\nPerformance results written to {out_path}")
+    if sink and run_dir is not None:
+        export_results(sink, run_dir)
 
 
 @pytest.fixture(autouse=True)
-def _register_sink(request, perf_sink, perf_run_dir):
-    """Inject sink and run_dir into the session for pytest_sessionfinish."""
+def _register_sink(request: pytest.FixtureRequest, perf_sink: list, perf_run_dir: Path) -> None:
     request.session._perf_sink = perf_sink  # type: ignore[attr-defined]
     request.session._perf_run_dir = perf_run_dir  # type: ignore[attr-defined]
